% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLDataFrame-methods.R
\name{filter.SQLDataFrame}
\alias{filter.SQLDataFrame}
\title{b[letters[10:15], ]
Return SQLDataFrame rows with matching conditions}
\usage{
\method{filter}{SQLDataFrame}(.data, ...)
}
\arguments{
\item{.data}{A SQLDataFrame object.}

\item{...}{Logical predicates defined in terms of the variables in
‘.data’. Multiple conditions are combined with ‘&’. Only rows
where the condition evaluates to ‘TRUE’ are kept. See
\code{?dplyr::filter} for more details.}
}
\value{
A SQLDataFrame object with subset rows of the input
    SQLDataFrame object matching conditions.
}
\description{
Use ‘filter()’ to choose rows/cases where conditions
    are true.
}
\examples{
dbfile <- system.file("extdata/test.db", package = "SQLDataFrame")
obj <- SQLDataFrame(dbname = dbfile, dbtable = "state", dbkey = "state")
obj1 <- obj \%>\% filter(region == "West" & size == "medium")
obj1
}
