% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSQLDataFrame.R
\name{makeSQLDataFrame}
\alias{makeSQLDataFrame}
\title{Construct SQLDataFrame from in-memory data frame or on-disk file.}
\usage{
makeSQLDataFrame(filename, dbtable = NULL, dbkey = character(), conn,
  host, user, dbname = NULL, password = NULL, billing = character(0),
  type = c("SQLite", "MySQL", "BigQuery"), overwrite = FALSE,
  sep = ",", index = FALSE, ...)
}
\arguments{
\item{filename}{A \code{data.frame} or \code{DataFrame} object, or
a character string of the filepath to the text file that to be
saved as SQL database table. For filepath, the data columns
should not be quoted on disk.}

\item{dbtable}{A character string for the to-be-saved database
table name. If not provided, will use the name of the input
\code{data.frame} or \code{DataFrame} object, or the
\code{basename(filename)} without extension if \code{filename}
is a character string.}

\item{dbkey}{A character vector for the name of key columns that
could uniquely identify each row of the database table.}

\item{conn}{a valid \code{DBIConnection} from \code{SQLite},
\code{MySQL} or \code{BigQuery}. If provided, arguments of
`user`, `host`, `dbname`, `password` will be ignored.}

\item{host}{host name for MySQL database or project name for
BigQuery.}

\item{user}{user name for SQL database.}

\item{dbname}{database name for SQL connection.}

\item{password}{password for SQL database connection.}

\item{billing}{the Google Cloud project name with authorized
billing information. Will be required if user does not have write
permission to the provided \code{BigQueryConnection} in
\code{conn}. e.g., when the connection is a public data set.}

\item{type}{The SQL database type, supports "SQLite", "MySQL" and
"BigQuery".}

\item{overwrite}{Whether to overwrite the \code{dbtable} if already
exists. Default is FALSE.}

\item{sep}{a character string to separate the terms.  Not
‘NA_character_’. Default is \code{,}.}

\item{index}{Whether to create an index table. Default is FALSE.}

\item{...}{additional arguments to be passed.}
}
\value{
A \code{SQLDataFrame} object.
}
\description{
Given a file name, \code{makeSQLDataFrame} will write
    the file contents into SQL database, and open the database
    table as SQLDataFrame.
}
\details{
The provided file must has one or more columns to unique
    identify each row (no duplicate rows allowed). The file must be
    rectangular without rownames. (if rownames are needed, save it
    as a column.)
}
\examples{
mtc <- tibble::rownames_to_column(mtcars)

## data.frame input
obj <- makeSQLDataFrame(mtc, dbkey = "rowname")
obj

## character input
filename <- file.path(tempdir(), "mtc.csv")
write.csv(mtc, file= filename, row.names = FALSE, quote = FALSE)
obj <- makeSQLDataFrame(filename, dbkey = "rowname")
obj

## save as MySQL database
\dontrun{
localConn <- DBI::dbConnect(dbDriver("MySQL"),
                            host = "",
                            user = "",
                            password = "",
                            dbname = "")
makeSQLDataFrame(filename, dbtable = "mtcMysql", dbkey = "rowname", conn = localConn)
}
}
