% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindSQLDataFrame.R
\docType{methods}
\name{rbindUniq}
\alias{rbindUniq}
\alias{rbindUniq,SQLDataFrame-method}
\title{unique rbind of \code{SQLDataFrame} objects}
\usage{
\S4method{rbindUniq}{SQLDataFrame}(..., deparse.level = 1)
}
\arguments{
\item{...}{One or more \code{SQLDataFrame} objects. These can be
given as named arguments.}

\item{deparse.level}{See ‘?base::cbind’ for a description of this
argument.}
}
\value{
A \code{SQLDataFrame} object.
}
\description{
Performs unique rbind on \code{SQLDataFrame}
    objects. The result will be sorted by the \code{dbkey} columns.
}
\details{
\code{rbindUnique} supports aggregation of SQLDataFrame
    objects.
}
\examples{
test.db <- system.file("extdata/test.db", package = "SQLDataFrame")
con <- DBI::dbConnect(DBI::dbDriver("SQLite"), dbname = test.db)
obj <- SQLDataFrame(conn = con,
                    dbtable = "state",
                    dbkey = c("region", "population"))
obj_sub1 <- obj[1:10, 2:3]
obj_sub2 <- obj[8:15, 2:3]

## union
res_rbind <- rbindUniq(obj_sub1, obj_sub2)  ## sorted
res_rbind
dim(res_rbind)
}
