% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionSQLDataFrame.R
\docType{methods}
\name{union}
\alias{union}
\alias{union,SQLDataFrame,SQLDataFrame-method}
\title{Union of \code{SQLDataFrame} objects}
\usage{
\S4method{union}{SQLDataFrame,SQLDataFrame}(x, y, copy = FALSE)
}
\arguments{
\item{x}{A \code{SQLDataFrame} object.}

\item{y}{A \code{SQLDataFrame} object.}

\item{copy}{Whether to}
}
\value{
A \code{SQLDataFrame} object.
}
\description{
Performs union operations on \code{SQLDataFrame}
    objects.
}
\examples{
db1 <- system.file("extdata/test.db", package = "SQLDataFrame")
db2 <- system.file("extdata/test1.db", package = "SQLDataFrame")
obj1 <- SQLDataFrame(dbname = db1,
                     dbtable = "state",
                     dbkey = c("region", "population"))
obj2 <- SQLDataFrame(dbname = db2,
                     dbtable = "state1",
                     dbkey = c("region", "population"))
obj1_sub <- obj1[1:10, 2:3]
obj2_sub <- obj2[8:15,2:3]

## union
res_union <- union(obj1_sub, obj2_sub)  ## sorted
dim(res_union)

}
