% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSQLDataFrame.R
\name{join}
\alias{join}
\alias{left_join.SQLDataFrame}
\alias{inner_join.SQLDataFrame}
\alias{semi_join.SQLDataFrame}
\alias{anti_join.SQLDataFrame}
\title{join \code{SQLDataFrame} together}
\usage{
\method{left_join}{SQLDataFrame}(x, y, by = NULL, suffix = c(".x",
  ".y"), ...)

\method{inner_join}{SQLDataFrame}(x, y, by = NULL, suffix = c(".x",
  ".y"), ...)

\method{semi_join}{SQLDataFrame}(x, y, by = NULL, suffix = c(".x",
  ".y"), ...)

\method{anti_join}{SQLDataFrame}(x, y, by = NULL, suffix = c(".x",
  ".y"), ...)
}
\arguments{
\item{x}{\code{SQLDataFrame} objects to join.}

\item{y}{\code{SQLDataFrame} objects to join.}

\item{by}{A character vector of variables to join by.  If ‘NULL’,
the default, ‘*_join()’ will do a natural join, using all
variables with common names across the two tables. See
\code{?dplyr::join} for details.}

\item{suffix}{A character vector of length 2 specify the suffixes
to be added if there are non-joined duplicate variables in ‘x’
and ‘y’. Default values are ".x" and ".y".See
\code{?dplyr::join} for details.}

\item{...}{additional arguments to be passed.}
}
\value{
A \code{SQLDataFrame} object.
}
\description{
*_join functions for \code{SQLDataFrame} objects.
}
\examples{
db1 <- system.file("extdata/test.db", package = "SQLDataFrame")
db2 <- system.file("extdata/test1.db", package = "SQLDataFrame")
obj1 <- SQLDataFrame(dbname = db1,
                     dbtable = "state",
                     dbkey = c("region", "population"))
obj2 <- SQLDataFrame(dbname = db2,
                     dbtable = "state1",
                     dbkey = c("region", "population"))

obj1_sub <- obj1[1:10, 1:2]
obj2_sub <- obj2[8:15, 2:3]

left_join(obj1_sub, obj2_sub)
inner_join(obj1_sub, obj2_sub)
semi_join(obj1_sub, obj2_sub)
anti_join(obj1_sub, obj2_sub)
}
