% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSQLDataFrame.R
\name{makeSQLDataFrame}
\alias{makeSQLDataFrame}
\title{Construct SQLDataFrame from file.}
\usage{
makeSQLDataFrame(filename, dbkey = character(), dbname = NULL,
  dbtable = NULL, overwrite = FALSE, sep = ",", index = TRUE, ...)
}
\arguments{
\item{filename}{A \code{data.frame} or \code{DataFrame} object, or
a character string of the filepath to the text file that to be
saved as SQL database table.}

\item{dbkey}{A character vector of column name(s) that could
uniquely identify each row of the filename. Must be provided in
order to construct a SQLDataFrame.}

\item{dbname}{A character string of the filepath of to-be-saved
database file. If not provided, will use a
\code{tempfile(fileext = ".db")}.}

\item{dbtable}{A character string for the to be saved database
table name. If not provided, will use the name of the input
\code{data.frame} or \code{DataFrame} object, or the
\code{basename(filename)} without extension if \code{filename}
is a character string.}

\item{overwrite}{Whether to overwrite the \code{dbtable} if already
exists. Default is FALSE.}

\item{sep}{a character string to separate the terms.  Not
‘NA_character_’. Default is \code{,}.}
}
\description{
Given a file name, \code{makeSQLDataFrame} will write
    the file contents into SQLite database, and open the database
    table as SQLDataFrame.
}
\details{
The provided file must has one or more columns to unique
    identify each row (no duplicate rows allowed). The file must be
    rectangular without rownames. (if rownames are needed, save it
    as a column.)
}
\examples{
mtc <- tibble::rownames_to_column(mtcars)

## data.frame input
obj <- makeSQLDataFrame(mtc, dbkey = "rowname")
obj

## character input
filename <- file.path(tempdir(), "mtc.csv")
write.csv(mtc, file= filename, row.names = FALSE)
obj <- makeSQLDataFrame(filename, dbkey = "rowname")
obj

}
