% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindSQLDataFrame.R
\docType{methods}
\name{rbind}
\alias{rbind}
\alias{rbind,SQLDataFrame-method}
\title{rbind of \code{SQLDataFrame} objects}
\usage{
\S4method{rbind}{SQLDataFrame}(..., deparse.level = 1)
}
\arguments{
\item{...}{One or more \code{SQLDataFrame} objects. These can be
given as named arguments.}

\item{deparse.level}{See ‘?base::cbind’ for a description of this
argument.}
}
\value{
A \code{SQLDataFrame} object.
}
\description{
Performs rbind on \code{SQLDataFrame} objects.
}
\examples{
obj1 <- SQLDataFrame(dbname = "inst/extdata/test.db",
                     dbtable = "state",
                     dbkey = c("region", "population"))
obj2 <- SQLDataFrame(dbname = "inst/extdata/test1.db",
                     dbtable = "state1",
                     dbkey = c("region", "population"))
obj1_sub <- obj1[1:10, 2:3]
obj2_sub <- obj2[8:15,2:3]

## rbind
res_rbind <- rbind(obj1_sub, obj2_sub)
dim(res_rbind)
}
