% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSQLDataFrame.R
\name{saveSQLDataFrame}
\alias{saveSQLDataFrame}
\title{Save SQLDataFrame object as a new database table.}
\usage{
saveSQLDataFrame(x, dbname = tempfile(fileext = ".db"),
  dbtable = deparse(substitute(x)), overwrite = FALSE, index = TRUE,
  ...)
}
\arguments{
\item{x}{The \code{SQLDataFrame} object to be saved.}

\item{dbname}{A character string of the file path of to be saved
database file.}

\item{dbtable}{A character string for the to be saved database
table name. Default is the name of the input
\code{SQLDataFrame}.}

\item{overwrite}{Whether to overwrite the \code{dbtable} if already
exists. Default is FALSE.}

\item{index}{Whether to create the database index. Default is TRUE.}

\item{...}{other parameters passed to methods.}
}
\description{
The function to save \code{SQLDataFrame} object as a
    database table with a supplied path to database.
}
\examples{
dbname <- system.file("extdata/test.db", package = "SQLDataFrame")
obj <- SQLDataFrame(dbname = dbname, dbtable = "state", dbkey = "state")
obj1 <- obj[1:10, 2:3]
obj1 <- saveSQLDataFrame(obj1, dbtable = "obj_subset")
dbname(obj1)
dbtable(obj1)
}
