% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSQLDataFrame.R
\name{saveSQLDataFrame}
\alias{saveSQLDataFrame}
\title{Save SQLDataFrame object as a new database table.}
\usage{
saveSQLDataFrame(x, dbname = tempfile(fileext = ".db"),
  dbtable = deparse(substitute(x)), overwrite = FALSE, ...)
}
\arguments{
\item{x}{The \code{SQLDataFrame} object to be saved.}

\item{dbname}{A database file path to save the \code{SQLDataFrame}
object.}

\item{dbtable}{The name of the new database table.}

\item{...}{other parameters passed to methods.}
}
\description{
The function to save \code{SQLDataFrame} object as a
    database table with a supplied path to database.
}
\examples{
dbname <- system.file("extdata/test.db", package = "SQLDataFrame")
ss <- SQLDataFrame(dbname = dbname, dbtable = "state", dbkey = "state")
ss1 <- ss[1:10, 2:3]
saveSQLDataFrame(ss1, tempfile(fileext = ".db"))
}
