% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLDataFrame-class.R
\docType{methods}
\name{SQLDataFrame}
\alias{SQLDataFrame}
\alias{class:SQLDataFrame}
\alias{SQLDataFrame-class}
\alias{dbname,SQLDataFrame-method}
\alias{dbname}
\alias{dbtable,SQLDataFrame-method}
\alias{dbtable}
\alias{dbkey,SQLDataFrame-method}
\alias{dbkey}
\alias{"dbkey<-"}
\alias{dbkey<-}
\alias{dbkey<-,SQLDataFrame-method}
\alias{dbconcatKey,SQLDataFrame-method}
\alias{dbconcatKey}
\alias{ROWNAMES,SQLDataFrame-method}
\alias{ROWNAMES}
\alias{show,SQLDataFrame-method}
\alias{coerce}
\alias{coerce,SQLDataFrame,data.frame-method}
\alias{as.data.frame,SQLDataFrame-method}
\alias{coerce,SQLDataFrame,DataFrame-method}
\alias{coerce,data.frame,SQLDataFrame-method}
\alias{coerce,DataFrame,SQLDataFrame-method}
\alias{coerce,ANY,SQLDataFrame-method}
\title{SQLDataFrame class}
\usage{
SQLDataFrame(dbname = character(0), dbtable = character(0),
  dbkey = character(0), col.names = NULL)

\S4method{dbname}{SQLDataFrame}(x)

\S4method{dbtable}{SQLDataFrame}(x)

\S4method{dbkey}{SQLDataFrame}(x)

\S4method{dbkey}{SQLDataFrame}(x) <- value

\S4method{dbconcatKey}{SQLDataFrame}(x)

\S4method{ROWNAMES}{SQLDataFrame}(x)

\S4method{show}{SQLDataFrame}(object)

\S4method{as.data.frame}{SQLDataFrame}(x, row.names = NULL,
  optional = NULL, ...)
}
\arguments{
\item{dbname}{A character string for the database file path.}

\item{dbtable}{A character string for the table name in that
database. If not provided and there is only one table
available, it will be read in by default.}

\item{dbkey}{A character vector for the name of key columns that
could uniquely identify each row of the database table.}

\item{col.names}{A character vector specifying the column names you
want to read into the \code{SQLDataFrame}.}

\item{x}{An \code{SQLDataFrame} object}

\item{value}{The column name to be used as \code{dbkey(x)}}

\item{object}{An \code{SQLDataFrame} object.}

\item{row.names}{\code{NULL} or a character vector giving the row
names for the data frame. Only including this argument for the
\code{as.data.frame} generic. Does not apply to
\code{SQLDataFrame}. See \code{base::as.data.frame} for
details.}

\item{optional}{logical. If \code{TRUE}, setting row names and
converting column names is optional. Only including this
argument for the \code{as.data.frame} generic. Does not apply
to \code{SQLDataFrame}. See \code{base::as.data.frame} for
details.}

\item{...}{additional arguments to be passed.}
}
\value{
A \code{SQLDataFrame} object.
}
\description{
\code{SQLDataFrame} constructor, slot getters, show
    method and coercion methods to \code{DataFrame} and
    \code{data.frame} objects.
}
\examples{

## constructor
dbname <- system.file("extdata/test.db", package = "SQLDataFrame")
obj <- SQLDataFrame(dbname = dbname, dbtable = "state",
                    dbkey = "state")
obj
obj1 <- SQLDataFrame(dbname = dbname, dbtable = "state",
                     dbkey = c("region", "population"))
obj1

## slot accessors
dbname(obj)
dbtable(obj)
dbkey(obj)
dbkey(obj1)

dbconcatKey(obj)
dbconcatKey(obj1)

## ROWNAMES
ROWNAMES(obj[sample(10, 5), ])
ROWNAMES(obj1[sample(10, 5), ])

## coercion
as.data.frame(obj)
as(obj, "DataFrame")


## dbkey replacement
dbkey(obj) <- c("region", "population")
obj
}
